+++ src/DotNetWorker/Configuration/ServiceCollectionExtensions.cs
@@ -81,7 +81,8 @@ namespace Microsoft.Extensions.DependencyInjection
 
         internal static IServiceCollection RegisterDefaultConverters(this IServiceCollection services)
         {
-            return services.AddSingleton<IConverter, OutputBindingConverter>()
+            return services.AddSingleton<IConverter, FunctionContextConverter>()
+                           .AddSingleton<IConverter, OutputBindingConverter>()
                            .AddSingleton<IConverter, TypeConverter>()
                            .AddSingleton<IConverter, JsonPocoConverter>();
         }
