+++ src/DotNetWorker/Converters/FunctionContextConverter.cs
@@ -1,0 +1,19 @@
+﻿namespace Microsoft.Azure.Functions.Worker.Converters
+{
+    internal class FunctionContextConverter : IConverter
+    {
+        public bool TryConvert(ConverterContext context, out object? target)
+        {
+            target = null;
+
+            // Special handling for the context.
+            if (context.Parameter.Type == typeof(FunctionExecutionContext))
+            {
+                target = context.ExecutionContext;
+                return true;
+            }
+
+            return false;
+        }
+    }
+}
