+++ src/DotNetWorker/Grpc/GrpcFunctionsHostLogger.cs
