+++ src/DotNetWorker/Grpc/GrpcFunctionsHostLoggerProvider.cs
